Added functionality to assign a bookable date as dropin.
Calendar now displays tuesdays as touch up days and takes into account if a day is a dropin day. These days are displayed in specific colors.

Added entity that should handle customizable text for the web page.

Added card that takes in text to the front page. This should take in text from entity eventually.