Start hour for dropin changed to 11.
Added calendar to model if from date is after to date in /add-dates/confirm-dates endpoint.

added calendar parameters and calendar to /booking/book-tattoo endpoint, to make selection of date to book easier.

send-bookable-date-through-calendar.js added to enable selection of date to book from calendar

Changed class to classappend and attribute set to date in calendar-layout, to enable selection functionality in booking.