Changed BookableDateForCalendarDto hours variable to List<String> for easier access in Javascript.
BookableDateService now converts a BookableHourForCalendarDto to a String, to comply with above.

show-hours-tooltip-on-date-mouseover.js introduced to show available hours when hovering the cursor over a date, to show booking information to customers.

styling added to style.css for tooltip.

calendar-layout now imports show-hours-tooltip-on-date-mouseover.js to show available hours on mouseover