Bookings at current date added to book-tattoo-with-date.html, for overview reasons when booking a customer.
getBookableDateByDate in BookableDateService now sorts the hours in a bookabledate.

BookingController now adds bookingsAtDate for use in template book-tattoo-with-date.html

getBookingsByDate in BookingService now takes in a date and returns bookings in the span from start of date to end of date.

bottom-columns class in calendar-layout.html now renamed calendar-bottom-columns