Customer handling is now done in CustomerController.java.
You can now search for a specific customer, view their bookings and delete the customer in the search-customer endpoint.

Deleting a customer now deletes the images assigned to their bookings and deletes sets the bookings customer to null if the booking is in the past. If it is in the future, it cancels the appointment.