Touch up start hours is no longer static from env variables.
Added javascript to handle checking the touch up radio button and replacing the time checkboxes with a time input.

Saving touch ups in the service now uses the same logic as other booking types.

Added check to not mark a date as booked if the booking is on a touch up date.